Implement axes reordering for input data
Depending on the models used (e.g CNN or ViT based), there are
different requirements on the order of channels in input data.

Add feature to dynamically read the original channel orders of the
data and the expect channel orders based on the selected models from
config files and adjust accordingly.

The `MultimodalDataModule` implements a `transform` function with
proper logic based on the config files. The `transform` function is
passed to the `DatasetHandler` upon instantiation who performs the
transformation.

Signed-off-by: Saeed Gholami Shahbandi <saeed.gh.sh@gmail.com>